split radio question from block and response
