+++ apps/api-journeys/src/application/application.ts
@@ -1,6 +1,6 @@
 import { createApplication } from 'graphql-modules'
 import { schemaBuilder } from '@core/shared/util-graphql'
-import { blockModule, journeyModule, responseModule } from '..'
+import { blockModule, journeyModule, responseModule } from '../modules'
 
 export const application = createApplication({
   modules: [blockModule, journeyModule, responseModule],
