+++ apps/api-journeys/src/modules/block/block.spec.ts
@@ -233,190 +233,6 @@ describe('BlockModule', () => {
     ])
   })
 
-  it('returns RadioOptionBlock', async () => {
-    const radioQuestionId = uuidv4()
-    const blockId = uuidv4()
-    const radioOption1: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'RadioOptionBlock',
-      parentBlockId: radioQuestionId,
-      parentOrder: 3,
-      extraAttrs: {
-        label: 'label',
-        description: 'description',
-        action: {
-          gtmEventName: 'gtmEventName',
-          blockId
-        }
-      }
-    }
-    const radioOption2: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'RadioOptionBlock',
-      parentBlockId: radioQuestionId,
-      parentOrder: 4,
-      extraAttrs: {
-        label: 'label',
-        description: 'description',
-        action: {
-          gtmEventName: 'gtmEventName',
-          journeyId
-        }
-      }
-    }
-    const radioOption3: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'RadioOptionBlock',
-      parentBlockId: radioQuestionId,
-      parentOrder: 5,
-      extraAttrs: {
-        label: 'label',
-        description: 'description',
-        action: {
-          gtmEventName: 'gtmEventName',
-          url: 'https://jesusfilm.org'
-        }
-      }
-    }
-    const radioOption4: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'RadioOptionBlock',
-      parentBlockId: radioQuestionId,
-      parentOrder: 6,
-      extraAttrs: {
-        label: 'label',
-        description: 'description',
-        action: {
-          gtmEventName: 'gtmEventName'
-        }
-      }
-    }
-    dbMock.block.findMany.mockResolvedValue([
-      radioOption1,
-      radioOption2,
-      radioOption3,
-      radioOption4
-    ])
-    const { data } = await query(gql`
-      query ($id: ID!) {
-        journey(id: $id) {
-          blocks {
-            id
-            __typename
-            parentBlockId
-            ... on RadioOptionBlock {
-              label
-              action {
-                __typename
-                gtmEventName
-                ... on NavigateToBlockAction {
-                  blockId
-                }
-                ... on NavigateToJourneyAction {
-                  journeyId
-                }
-                ... on LinkAction {
-                  url
-                  target
-                }
-              }
-            }
-          }
-        }
-      }
-    `)
-    expect(data?.journey.blocks).toEqual([
-      {
-        id: radioOption1.id,
-        __typename: 'RadioOptionBlock',
-        parentBlockId: radioQuestionId,
-        label: 'label',
-        action: {
-          __typename: 'NavigateToBlockAction',
-          gtmEventName: 'gtmEventName',
-          blockId
-        }
-      },
-      {
-        id: radioOption2.id,
-        __typename: 'RadioOptionBlock',
-        parentBlockId: radioQuestionId,
-        label: 'label',
-        action: {
-          __typename: 'NavigateToJourneyAction',
-          gtmEventName: 'gtmEventName',
-          journeyId
-        }
-      },
-      {
-        id: radioOption3.id,
-        __typename: 'RadioOptionBlock',
-        parentBlockId: radioQuestionId,
-        label: 'label',
-        action: {
-          __typename: 'LinkAction',
-          gtmEventName: 'gtmEventName',
-          url: 'https://jesusfilm.org',
-          target: null
-        }
-      },
-      {
-        id: radioOption4.id,
-        __typename: 'RadioOptionBlock',
-        parentBlockId: radioQuestionId,
-        label: 'label',
-        action: {
-          __typename: 'NavigateAction',
-          gtmEventName: 'gtmEventName'
-        }
-      }
-    ])
-  })
-
-  it('returns RadioQuestionBlock', async () => {
-    const parentBlockId = uuidv4()
-    const radioQuestion: Block = {
-      id: uuidv4(),
-      journeyId,
-      blockType: 'RadioQuestionBlock',
-      parentBlockId,
-      parentOrder: 2,
-      extraAttrs: {
-        label: 'label',
-        description: 'description'
-      }
-    }
-    dbMock.block.findMany.mockResolvedValue([radioQuestion])
-    const { data } = await query(gql`
-      query ($id: ID!) {
-        journey(id: $id) {
-          blocks {
-            id
-            __typename
-            parentBlockId
-            ... on RadioQuestionBlock {
-              label
-              description
-            }
-          }
-        }
-      }
-    `)
-    expect(data?.journey.blocks).toEqual([
-      {
-        id: radioQuestion.id,
-        __typename: 'RadioQuestionBlock',
-        parentBlockId,
-        label: 'label',
-        description: 'description'
-      }
-    ])
-  })
-
   it('returns StepBlock', async () => {
     const parentBlockId = uuidv4()
     const nextBlockId = uuidv4()
