+++ apps/api-journeys/src/modules/response/response.spec.ts
@@ -17,95 +17,6 @@ describe('Response', () => {
     })
   })
 
-  describe('radioQuestionResponseCreate', () => {
-    it('creates a radio question block response', async () => {
-      const userId = uuidv4()
-      const block1: Block = {
-        id: uuidv4(),
-        journeyId: uuidv4(),
-        blockType: 'RadioQuestionBlock',
-        parentBlockId: null,
-        parentOrder: 0,
-        extraAttrs: {
-          label: 'label',
-          description: 'description'
-        }
-      }
-      dbMock.block.findUnique.mockResolvedValue(block1)
-      const response1: Response = {
-        id: uuidv4(),
-        type: 'RadioQuestionResponse',
-        blockId: block1.id,
-        userId,
-        extraAttrs: {
-          radioOptionBlockId: uuidv4()
-        },
-        createdAt: new Date(),
-        updatedAt: new Date()
-      }
-      dbMock.response.create.mockResolvedValue(response1)
-      const { data } = await testkit.execute(app, {
-        document: gql`
-          mutation ($input: RadioQuestionResponseCreateInput!) {
-            radioQuestionResponseCreate(input: $input) {
-              id
-              userId
-              radioOptionBlockId
-              block {
-                id
-                label
-                description
-              }
-            }
-          }
-        `,
-        variableValues: {
-          input: {
-            id: response1.id,
-            blockId: response1.blockId,
-            radioOptionBlockId: get(response1.extraAttrs, 'radioOptionBlockId')
-          }
-        },
-        contextValue: {
-          db: dbMock,
-          userId
-        }
-      })
-      expect(data?.radioQuestionResponseCreate).toEqual({
-        id: response1.id,
-        userId,
-        radioOptionBlockId: get(response1.extraAttrs, 'radioOptionBlockId'),
-        block: {
-          id: block1.id,
-          label: 'label',
-          description: 'description'
-        }
-      })
-    })
-
-    it('throws authentication error if no user token', async () => {
-      const { errors } = await testkit.execute(app, {
-        document: gql`
-          mutation ($input: RadioQuestionResponseCreateInput!) {
-            radioQuestionResponseCreate(input: $input) {
-              id
-            }
-          }
-        `,
-        variableValues: {
-          input: {
-            blockId: uuidv4(),
-            radioOptionBlockId: uuidv4()
-          }
-        },
-        contextValue: {
-          db: dbMock
-        }
-      })
-      expect(errors?.[0].extensions?.code).toEqual('UNAUTHENTICATED')
-    })
-  })
-
   describe('videoResponseCreate', () => {
     it('creates a video block response', async () => {
       const userId = uuidv4()
